Prompts Opus to encapsulate fullscreen functionality into a hook
